Updated Permission Structure
Instread of using the previous hard-coded Permission Classses,
every permission is now an instance of the Base class Permission.

The class PermissionClasses is renamed to Permissions. It contains a
List of Permissions. Application level permissions are declared as a
static member of the Permissions class.

The setup class was also modified to comply with the changes.
Custom permission classes were deleted in order to reduce code.